Fixed player movement bugs
moved player coordinate change to after move function (player position in room change) because having it before cause a bug where the player would be put in the wrong place.
Removed + x, + y from statement in safeMove as it meant it was trying to put the underPlayer entity in the place the player was moved to, therefor removing the player.
Changed Exception to IndexOutOfBoundsException in roomInteract()